+++ wxc_sdk/telephony/forwarding.py
@@ -1,7 +1,7 @@
 """
 Forwarding settings and API for call queues, hunt groups, and auto attendants
 """
-
+import json
 from typing import Optional, Dict
 
 from pydantic import Field, validator
@@ -60,7 +60,7 @@ class ForwardingSetting(ApiModel):
 class CallForwarding(ApiModel):
     always: ForwardingSetting
     selective: ForwardingSetting
-    rules: list[ForwardingRule]
+    rules: Optional[list[ForwardingRule]]
 
     @staticmethod
     def default() -> 'CallForwarding':
@@ -278,18 +278,11 @@ class ForwardingApi:
         """
         params = org_id and {'orgId': org_id} or {}
         url = self._endpoint(location_id=location_id, feature_id=feature_id)
-        body = forwarding.dict()
 
-        # update only has 'id' and 'enabled' in rules
-        # determine names of ForwardingRule fields to remove
-        to_pop = [field
-                  for field in ForwardingRule.__fields__
-                  if field not in {'id', 'enabled'}]
-        for rule in body['rules']:
-            rule: Dict
-            for field in to_pop:
-                rule.pop(field, None)
-        body = {'callForwarding': body}
+        body = {'callForwarding': json.loads(forwarding.json(exclude={'rules': {'__all__': {'calls_from',
+                                                                                            'forward_to',
+                                                                                            'calls_to',
+                                                                                            'name'}}}))}
         self._session.rest_put(url=url, json=body, params=params)
 
     def create_call_forwarding_rule(self, location_id: str, feature_id: str,
@@ -314,9 +307,9 @@ class ForwardingApi:
         :rtype; str
         """
         url = self._endpoint(location_id=location_id, feature_id=feature_id, path='selectiveRules')
-        body = forwarding_rule.dict()
         params = org_id and {'orgId': org_id} or None
-        data = self._session.rest_post(url=url, json=body, params=params)
+        body = forwarding_rule.json()
+        data = self._session.rest_post(url=url, data=body, params=params)
         return data['id']
 
     def call_forwarding_rule(self, location_id: str, feature_id: str, rule_id: str,
@@ -374,8 +367,8 @@ class ForwardingApi:
         """
         url = self._endpoint(location_id=location_id, feature_id=feature_id, path=f'selectiveRules/{rule_id}')
         params = org_id and {'orgId': org_id} or None
-        body = forwarding_rule.dict()
-        data = self._session.rest_put(url=url, params=params, json=body)
+        body = forwarding_rule.json(exclude={'id'})
+        data = self._session.rest_put(url=url, params=params, data=body)
         return data['id']
 
     def delete_call_forwarding_rule(self, location_id: str, feature_id: str, rule_id: str, org_id: str = None):
@@ -391,4 +384,4 @@ class ForwardingApi:
         """
         url = self._endpoint(location_id=location_id, feature_id=feature_id, path=f'selectiveRules/{rule_id}')
         params = org_id and {'orgId': org_id} or None
-        self._session.delete(url=url, params=params)
+        self._session.rest_delete(url=url, params=params)